Split shell instance lease tests
